Update booster
- Update setup script
- Add method for updating swap min amount
- Upgrade tests to the latest fq version
- Fix passport required top up
- Fix account exceeding balance
- Add account and passport settings update on token deposit
- token_processing -> auto_reinvestment

+++ contracts/booster/interfaces/IBoosterPassport.sol
@@ -78,6 +78,6 @@ interface IBoosterPassport {
     event PingFrequencyUpdated(address account, uint128 frequency);
     event PingMaxPriceUpdated(uint128 price);
     event AutoPingUpdated(address account, bool status);
-    event Ping(address account, uint _timestamp, uint counter, bool byManager);
+    event Ping(address account, uint128 _timestamp, uint counter, bool byManager);
     event ManagersUpdated(uint[] managers);
 }
